Configure basic environment settings
The dev environment has been configured to use port 1234 by default (as
I expect the rails API to run on port 3000).

I've also added a basic template for the environments file. Anyone who
runs this app will either have to add the missing variable to this file
or create a new `.env*.local` file that includes these environment
variables. The rewritten endpoints used in the frontend's proxy (for
requests made to the API) depends on the environment variables set in
these files.