Add the WorldMap components
This commit adds a basic WorldMap component to the index page. This will
be used to show the geolocation of different page visits but at present
it will only display a marker for the city of Buenos Aires.
This commit also introduces the DateInput component which will be used
to set the date range for the visits displayed.