Add support for react-simple-maps
I've had to fork the react-simple-maps package because a lot of its
dependencies were so out of date. I refactored the package significantly
because it was originally designed to support CJS as well as ESM but for
some reason the version of the package that this repo kept importing was
the CJS version which would raise errors (`[ERR_REQUIRE_ESM]: require()
of ES Module not supported`). My fork [1] only supports ESM and uses
`.jsx` files (rather than `.js`) for each of the components.

[1]: https://github.com/kohrvid/react-simple-maps