Create the first "page" for visits
To simplify matters, I've decided to add a set of pages that will fire
off a request to the /api/visits endpoint when a user visits. For now
each user will be treated as an anonymous one though this will be
refactored in a later commit.
As it isn't possible to collect the IP address for page visits made to
localhost, I've decided to add ngrok as a requirement for this project.
An ngrok tunnell will allow users to access the page from a publicly
accessible URL whilst in the development environment, meaning that the
`X-FORWARDED-FOR` header will capture a user's public IP address.

This commit also fixes a bug in the NextJS config that would have
treated API functions as React components in error.