Add a pages dropdown to the frontend
This commit adds a new SelectInput components which is used to render
the pages dropdown on the index page. Its options are generated from the
response of the `GET /pages` endpoint on the backend.
The dropdown itself is needed to determine which page's metrics will be
displayed on the front page.