Increase width of ComposableMap and fix styling
This component defaults to 800 which I've noticed crops out half of New
Zealand when used with my GeoJSON map. 1200 seems to capture all of the
countries plotted.
I've also made a few changes to the colours used in the markers. Adding
some translucency in particular makes it easier to spot regions with
higher page traffic.