Rename DateInput to DateTimeInput
It makes more sense to allow users to select the date and time in this
component rather than just the date (particularly if one wanted to
monitor changes in page traffic over the course of a single day). As
such I've replaced the date picker with a datetime input.  I've also
refactored the `key` prop for the `option` elements so that it uses an
index instead of the option's ID. I expect that there could be instances
where the ID is null(?) so this might be more appropriate.