Allow users to filter visits by page ID and time
This commit refactors the fetch request for the list of visits and
renames the function `listVisits` now that it is possible to filter
visits by the `pageId`, `fromDate` and `toDate` state values. In order
to accommodate this, the pageId will now default to `null` rather than
`undefined` so that it is easier to determine whether this value should
be added to the query string or not. An extra useEffect has been added
so that all changes made to these values will trigger a fetch request to
the backend with the new parameters.

I've also added an initial "all pages" option to the dropdown list so
that users can easily remove this filter and view markers for all page
visits.