Set a currentUser in the application
This commit adds a new fetch request for the `/users/auth` endpoint on
the API which will be used to check if a user is logged in (via their
session cookie). When a user is logged in, the new `currentUser` state
value will be set with their username and ID and `authenticated` will be
set to true, allowing me to check a user's logged in status on different
parts of the frontend app. The SessionProvider will also check if an
admin user is logged in and set the `adminAuthenticated` state value
accordingly.