Set the userId when creating Visits
Following on from the previous commit, I can now set the user ID in the
`createVisit` request when users are logged in, allowing me to track
the visits made by individual users. Because this depends on a useEffect
(in fact two, one when authenticating the user in the SessionContext and
another when creating the visit in the VisitContext), the request made
will fire twice in the development environment [1], something that could
skew the statistics somewhat during testing and development.

[1]: https://react.dev/reference/react/useEffect#troubleshooting